feat: better fallback for Windows
This adds a better fallback approach for Windows and systems lacking the
`sleep` CLI utility. This relies on spawning a Node.js process designed
to sleep for the desired time.

This has a tradeoff that it typically sleeps for longer than the desired
time, but it's available on all systems and is more predictable than
`timeout` for Windows. This makes busy-waiting obsolete, so that section
and test case have been removed.

This should still avoid blocking the CPU, which is a more critical
concern.

Fixes #1, #2